requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
wheel==0.45.1
